requirements.txt
apache-airflow[kubernetes]==2.7.3
psycopg2-binary==2.9.7
kubernetes==26.1.0
